refactor: split api server routes
